Introduce `CELERY_BROKER_URL` and `CELERY_BACKEND_URL` as Env variables
A longstanding issue we have had in the Canso Dev Agent is that Celery broker and backend
URLs have been hardcoded in the application. Since we run 2 Dev agents in the data plane
(one for staging & prod each), the hardcoded Broker & Backend URLs cause the staging Dev
Agent to interact with the Prod celery & redis.

This caused issues when the Dev Agent refactor was being tested. We had to hardcode the URLs
to the staging values to then see the logs in the Dev Agent and Celery PODs.

This PR makes a longer term fix so that we can manage both deployments of Dev Agents in the
Data Plane independently. In both the Celery Worker & Dev Agent Deployment, we're adding
CELERY_BROKER_URL and CELERY_BACKEND_URL as env variables. The logic to get these values in
a reliable fashion uses Go's templating functionality and is explained below

- Overall format of the value is `"redis://%s.%s.svc.cluster.local:%d/%d"`
- The 1st element is the service name, which we get from the helper `canso-agent-celery.fullname`.
This has been defined already in _helpers and is used by the Redis Service as well. Using the same
helper ensures alignment.
- The 2nd element is the namespace, which is taken from the Helm chart's release
- The 3rd element is the Redis Service port. Now, this was hardcoded to 6379 in Redis template. To
ensure a single source of truth, a `cansoAgent.redis` now has a param `servicePort`. This is now
used in the Redis Service as well as as the 3rd element, thereby ensuring alignment
- The 4th value is 0 in all cases we've seen. It can be gotten from `.Values.cansoAgent.redis.database`
and defaults to 0. To keep things simple, haven't added `database` in values.yaml, so the default 0
will get picked always.
- servicePort is cast as integer (3rd element)
- database is cast as integer (4th element)

Helm template commands were run to render the values and check if the result look good. The results will
be attached in the main PR for greater visibility.

Use Redis service name (incorrectly used Celery service name